+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Certs.hs
@@ -23,9 +23,12 @@ module Cardano.Ledger.Conway.Rules.Certs (
   ConwayCertsEvent (..),
   CertsEnv (..),
   updateDormantDRepExpiry,
+  updateDormantDRepExpiries,
+  updateVotingDRepExpiries,
 ) where
 
 import Cardano.Ledger.BaseTypes (
+  EpochInterval,
   EpochNo (EpochNo),
   Globals (..),
   ShelleyBase,
@@ -42,7 +45,12 @@ import Cardano.Ledger.Binary.Coders (
   (<!),
  )
 import Cardano.Ledger.Conway.Core
-import Cardano.Ledger.Conway.Era (ConwayCERT, ConwayCERTS, ConwayEra)
+import Cardano.Ledger.Conway.Era (
+  ConwayCERT,
+  ConwayCERTS,
+  ConwayEra,
+  hardforkConwayMoveWithdrawalsAndDRepChecksToLedgerRule,
+ )
 import Cardano.Ledger.Conway.Governance (
   Committee,
   GovActionPurpose (..),
@@ -104,7 +112,7 @@ deriving instance (EraPParams era, Show (Tx era)) => Show (CertsEnv era)
 instance (EraPParams era, NFData (Tx era)) => NFData (CertsEnv era)
 
 data ConwayCertsPredFailure era
-  = -- | Withdrawals that are missing or do not withdraw the entire amount
+  = -- | Withdrawals that are missing or do not withdraw the entire amount (pv < 11)
     WithdrawalsNotInRewardsCERTS Withdrawals
   | -- | CERT rule subtransition Failures
     CertFailure (PredicateFailure (EraRule "CERT" era))
@@ -183,6 +191,7 @@ instance
   , Embed (EraRule "CERT" era) (ConwayCERTS era)
   , EraCertState era
   , ConwayEraCertState era
+  , EraRuleFailure "CERT" era ~ PredicateFailure (EraRule "CERT" era)
   ) =>
   STS (ConwayCERTS era)
   where
@@ -204,6 +213,7 @@ conwayCertsTransition ::
   , Embed (EraRule "CERT" era) (ConwayCERTS era)
   , Environment (EraRule "CERT" era) ~ CertEnv era
   , Signal (EraRule "CERT" era) ~ TxCert era
+  , EraRuleFailure "CERT" era ~ PredicateFailure (EraRule "CERT" era)
   ) =>
   TransitionRule (ConwayCERTS era)
 conwayCertsTransition = do
@@ -213,58 +223,75 @@ conwayCertsTransition = do
       , certificates
       ) <-
     judgmentContext
-  network <- liftSTS $ asks networkId
-
   case certificates of
-    Empty -> do
-      let drepActivity = pp ^. ppDRepActivityL
-      -- If there is a new governance proposal to vote on in this transaction,
-      -- AND the number of dormant-epochs recorded is greater than zero, we bump
-      -- the expiry for all DReps by the number of dormant epochs, and reset the
-      -- counter to zero.
-      -- It does not matter that this rule (CERTS) is called _before_ the GOV rule
-      -- in LEDGER, even though we cannot validate any governance proposal here,
-      -- since the entire transaction will fail if the proposal is not accepted in
-      -- GOV, and so will this expiry bump done here. It will be discarded.
-      let certState' =
-            let hasProposals = not . OSet.null $ tx ^. bodyTxL . proposalProceduresTxBodyL
-             in if hasProposals
-                  then certState & certVStateL %~ updateDormantDRepExpiry currentEpoch
-                  else certState
-
-      -- Update DRep expiry for all DReps that are voting in this transaction.
-      -- This will execute in mutual-exclusion to the previous updates to DRep expiry,
-      -- because if there are no proposals to vote on , there will be no votes either.
-      let numDormantEpochs = certState' ^. certVStateL . vsNumDormantEpochsL
-          updateVSDReps vsDReps =
-            Map.foldlWithKey'
-              ( \dreps voter _ -> case voter of
-                  DRepVoter cred ->
-                    Map.adjust
-                      (drepExpiryL .~ computeDRepExpiry drepActivity currentEpoch numDormantEpochs)
-                      cred
-                      dreps
-                  _ -> dreps
-              )
-              vsDReps
-              (unVotingProcedures $ tx ^. bodyTxL . votingProceduresTxBodyL)
-
-      -- Final CertState with updates to DRep expiry based on new proposals and votes on existing proposals
-      let certStateWithDRepExpiryUpdated = certState' & certVStateL . vsDRepsL %~ updateVSDReps
-          dState = certStateWithDRepExpiryUpdated ^. certDStateL
-          withdrawals = tx ^. bodyTxL . withdrawalsTxBodyL
-
-      -- Validate withdrawals and rewards and drain withdrawals
-      failOnJust
-        (withdrawalsThatDoNotDrainAccounts withdrawals network (dState ^. accountsL))
-        WithdrawalsNotInRewardsCERTS
-
-      pure $ certStateWithDRepExpiryUpdated & certDStateL . accountsL %~ drainAccounts withdrawals
+    Empty ->
+      if hardforkConwayMoveWithdrawalsAndDRepChecksToLedgerRule $ pp ^. ppProtocolVersionL
+        then pure certState
+        else do
+          network <- liftSTS $ asks networkId
+          let accounts = certState ^. certDStateL . accountsL
+              withdrawals = tx ^. bodyTxL . withdrawalsTxBodyL
+          failOnJust
+            (withdrawalsThatDoNotDrainAccounts withdrawals network accounts)
+            ( \(invalid, incomplete) ->
+                WithdrawalsNotInRewardsCERTS $ Withdrawals $ unWithdrawals invalid <> unWithdrawals incomplete
+            )
+          pure $
+            certState
+              & updateDormantDRepExpiries tx currentEpoch
+              & updateVotingDRepExpiries tx currentEpoch (pp ^. ppDRepActivityL)
+              & certDStateL . accountsL %~ drainAccounts withdrawals
     gamma :|> txCert -> do
       certState' <-
         trans @(ConwayCERTS era) $ TRC (env, certState, gamma)
       trans @(EraRule "CERT" era) $
         TRC (CertEnv pp currentEpoch committee committeeProposals, certState', txCert)
+
+-- | If there is a new governance proposal to vote on in this transaction,
+-- AND the number of dormant-epochs recorded is greater than zero, we bump
+-- the expiry for all DReps by the number of dormant epochs, and reset the
+-- counter to zero.
+--
+-- It does not matter that this is called _before_ the GOV rule in LEDGER, even
+-- though we cannot validate any governance proposal here, since the entire
+-- transaction will fail if the proposal is not accepted in GOV, and so will
+-- this expiry bump done here.
+updateDormantDRepExpiries ::
+  ( EraTx era
+  , ConwayEraTxBody era
+  , ConwayEraCertState era
+  ) =>
+  Tx era -> EpochNo -> CertState era -> CertState era
+updateDormantDRepExpiries tx currentEpoch =
+  let hasProposals = not . OSet.null $ tx ^. bodyTxL . proposalProceduresTxBodyL
+   in if hasProposals
+        then certVStateL %~ updateDormantDRepExpiry currentEpoch
+        else id
+
+-- | Update DRep expiry for all DReps that are voting in this transaction. This
+-- will execute in mutual-exclusion to the updates to the dormant DRep expiry,
+-- because if there are no proposals to vote on, there will be no votes either.
+updateVotingDRepExpiries ::
+  ( EraTx era
+  , ConwayEraTxBody era
+  , ConwayEraCertState era
+  ) =>
+  Tx era -> EpochNo -> EpochInterval -> CertState era -> CertState era
+updateVotingDRepExpiries tx currentEpoch drepActivity certState =
+  let numDormantEpochs = certState ^. certVStateL . vsNumDormantEpochsL
+      updateVSDReps vsDReps =
+        Map.foldlWithKey'
+          ( \dreps voter _ -> case voter of
+              DRepVoter cred ->
+                Map.adjust
+                  (drepExpiryL .~ computeDRepExpiry drepActivity currentEpoch numDormantEpochs)
+                  cred
+                  dreps
+              _ -> dreps
+          )
+          vsDReps
+          (unVotingProcedures $ tx ^. bodyTxL . votingProceduresTxBodyL)
+   in certState & certVStateL . vsDRepsL %~ updateVSDReps
 
 instance
   ( Era era